Created classes.js
I tried a few ways of getting the typing to work. Some of those worked
but I had this idea of making classes so I looked it up and that worked.

script.js now imports the classes at the top of the file.
I think I need to clean some unused code now that I have the classes.
The ninety or so lines of coments in script.js were in some ways
successful in helping me figure out what I was doing.

The classes I first tried using a different repo. Copied from there,
they only needed a few changes to work here. Sentences are made of
Words, and Words are made of Chars. The active sentence loops through
the words and the active word loops through the chars. The active
Char returns a promise, it focuses on itself, IF the div was previously
marked incorrect, it rejects, removes it's 'current' class, and
returns an error message. A handler gets the keydown event and,
IF the key pressed matches the active Char, it resolves the promise,
adds the 'correct' class to the div, removes the handler(for good
measure), and moves on to the next Char. else IF the key pressed is
the spacebar, it marks the ramaining chars as incorrect and skipped,
removes the 'current' class, adds the 'skipped' class to the parent word
, and resolves. Else all other keys mark the Char as incorrect,
removes the 'current' class, and resolves.
The SpecChar class is there to handle spaces, punctuation, and other
non-letter characters. Right now, every word makes itself with a 'split'
which defaults to a space. It also handles its own 'isCurrent' method
which works similarly to the Char class.
The Word class when current waits for all the chars to resolve
before it resolves itself. IF the word becomes 'skipped', it removes
the 'current' class from itself and rejects. I gave it a catch because
I ran into some errors but I got things working before really
understanding how the catch works. That's to say that I'm still not
confident in how these async functions work together.
The Sentence class is a lot more simple.