Added function to track WPM and accuracy
first I made it so sentences are flex and wrap in their container.
I made a prompt class to handle any number of sentences.
The prompt handles the WPM by taking startTime when the isCurrent method
is called then endTime when the for loop is done.
The accuracy is calculated during each of the Chars isCurrent method so
it updates with each input.
I want to add backspace support and make new prompt subclasses.
I also want to work on the UI and I'm sure there's more functions to
figure out but I want to start thinking about the actual game experience