Began designing the UI & added float to prompts
Mostly small changes in style I still need to clean up that file.
Added 'fromString' to Sentence class to create a Sentence from a string.
I haven't tested it yet.
Improved the Prompt constructor, added the float method which allows the
prompt to call a given function when 'resolved'.
Made GameInput which is a form that handles user input when there are
multiple options.
In script I made it so the page opens with a 'scene0'
there the user can select an option by typing enough to match the prompt
then submit with enter.
I want to experiment with making scenes as functions in a seperate file