added two options to scene zero
They even seem to work as intended so far. I removed the buttons since
those functions can now be done by key inputs. That said, I have to
implement a pause key and I don't even know which key that should be.
I put the GameInput 'UserInput' in the 'DOMContentLoaded' function so
it seems that it can stay the same for the whole game. So as it is now,
displayScene0 will be like the home screen (I think it might have a
title, subtitle and prompts for start game, info, and options.) . The
two promps may be selected by typing the words into the input field
and pressing enter. pStartGame will play ten random words, then give the
option to reset. pTestSentence will play the test sentence, then give
the option to reset. Right now, pressing enter on an empty input
seems to select the first option. I don't know how I feel about that.
I found that there's an error flagged when words are skipped. I think
that's just because the rejection flags an error. I think I can use
that skipped condition for some kind of feature in the game.