I'm gonna start with making five prompts
Commit before I start implementing a module of functions.
I'm not perfectly satisfied with the CSS yet but I want to work on the
game logic and come back to the CSS later.
In this condition, the game will play the random sentence and even the
test sentence, but there were some errors in the console that I know to
keep an eye out for. I think I improved the classes and made things work
better overall. I added the 'Timer' class which each prompt can use to
keep track of time. I managed to give the PInput class the method to
filter an array of options. I know there's an error in the float method
of the Prompt class that I need to fix.