I like the current design of the css
Scifi.js works by importing the classes then making the sentences,
making prompts from those sentences, then making a function that runs
the a 'Page' of the game. the page runs the prompts one by one and
places each prompt inside a previousDiv after it is typed. After the
page is completed, the completePage function displays the options to
the user.

I found that I don't have to export every variable and function from
scifi to script. I wonder if that's a quirk of how modules work in js.

I made a handful of changes in classes and style including logic that
sets the div height for each prompt. and grid-template-areas that
control the screen layout.