Started the draculaDaily
The new Scene class seems to be working properly. The constructor takes
a gameContainer div, and an array of Prompt objects. When playScene() is
called, the scene changes the class of the gameContainer to fit the
prompt and runs each prompt before calling playOptions() to display the
Scene.options. I think the options for each scene should be set manually
when the scene is built. so I may change how the options are stored but
things are working for now.