Started the Maze.js
The scene class did work pretty well as expected. The maze can be
followed, as is to the exit. I'm gonna rework the Scene class now so
navigating scenes back and forth may be easier. Basically I want to make
every scene a const and feed the list of prompts to the playScene method
so the scene changes based on the prompts. The options should work
similarly. I don't know if this will create too many objects in memory.
Not sure if that's even a concern.